+++ data/src/main/java/com/microsoft/azure/kusto/data/ClientFactory.java
@@ -3,7 +3,9 @@
 
 package com.microsoft.azure.kusto.data;
 
+import com.azure.core.http.HttpClient;
 import com.microsoft.azure.kusto.data.auth.ConnectionStringBuilder;
+import com.microsoft.azure.kusto.data.http.HttpClientProperties;
 import com.microsoft.azure.kusto.data.http.HttpClientFactory;
 import org.apache.http.impl.client.CloseableHttpClient;
 
@@ -11,6 +13,7 @@ import java.net.URISyntaxException;
 import java.util.Optional;
 
 public class ClientFactory {
+
     private ClientFactory() {
         // Hide the default constructor, as this is a factory with static methods
     }
@@ -45,12 +48,12 @@ public class ClientFactory {
      * customized with the given properties.
      *
      * @param csb the connection string builder
-     * @param client CloseableHttpClient client. It will not be closed when {@link Client#close} is called.
+     * @param client HttpClient client.
      * @return a fully constructed {@linkplain Client} instance
      * @throws URISyntaxException if the cluster URL is invalid
      */
-    public static Client createClient(ConnectionStringBuilder csb, CloseableHttpClient client) throws URISyntaxException {
-        return client == null ? createClient(csb, (HttpClientProperties) null) : new ClientImpl(csb, client, true);
+    public static Client createClient(ConnectionStringBuilder csb, HttpClient client) throws URISyntaxException {
+        return client == null ? createClient(csb, (HttpClientProperties) null) : new ClientImpl(csb, client);
     }
 
     /**
@@ -75,11 +78,7 @@ public class ClientFactory {
      * @throws URISyntaxException if the cluster URL is invalid
      */
     public static StreamingClient createStreamingClient(ConnectionStringBuilder csb, HttpClientProperties properties) throws URISyntaxException {
-        if (properties == null) {
-            HttpClientProperties.builder().build();
-        }
-
-        return new ClientImpl(csb, HttpClientFactory.create(properties), false);
+        return new ClientImpl(csb, properties);
     }
 
     /**
@@ -87,11 +86,11 @@ public class ClientFactory {
      * is customized with the given properties.
      *
      * @param csb the connection string builder
-     * @param httpClient HTTP client
+     * @param httpClient HTTAP client
      * @return a fully constructed {@linkplain StreamingClient} instance
      * @throws URISyntaxException if the cluster URL is invalid
      */
-    public static StreamingClient createStreamingClient(ConnectionStringBuilder csb, CloseableHttpClient httpClient) throws URISyntaxException {
-        return new ClientImpl(csb, httpClient, true);
+    public static StreamingClient createStreamingClient(ConnectionStringBuilder csb, HttpClient httpClient) throws URISyntaxException {
+        return new ClientImpl(csb, httpClient);
     }
 }