+++ data/src/main/java/com/microsoft/azure/kusto/data/auth/CallbackTokenProvider.java
@@ -3,10 +3,10 @@
 
 package com.microsoft.azure.kusto.data.auth;
 
+import com.azure.core.http.HttpClient;
 import com.microsoft.azure.kusto.data.exceptions.DataClientException;
 
 import com.microsoft.azure.kusto.data.exceptions.ExceptionsUtils;
-import org.apache.http.client.HttpClient;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;