+++ data/src/main/java/com/microsoft/azure/kusto/data/auth/HttpClientWrapper.java
@@ -1,26 +1,14 @@
 package com.microsoft.azure.kusto.data.auth;
 
-import com.azure.core.http.HttpRequest;
-import com.azure.core.http.HttpResponse;
+import com.azure.core.http.*;
+import com.azure.core.util.Context;
+import com.azure.core.util.CoreUtils;
 import com.microsoft.aad.msal4j.IHttpClient;
 import com.microsoft.aad.msal4j.IHttpResponse;
-import org.apache.http.Header;
-import org.apache.http.HttpEntityEnclosingRequest;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.*;
-import org.apache.http.entity.ByteArrayEntity;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.InputStreamEntity;
-import org.apache.http.message.BasicHeader;
-import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
 
-import java.io.IOException;
-import java.io.PipedInputStream;
-import java.io.PipedOutputStream;
-import java.nio.charset.StandardCharsets;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
+import reactor.core.publisher.Mono;
+
+import java.util.stream.Collectors;
 
 /**
  * This class wraps both of the azure http client interfaces - IHttpClient and HttpClient to use our apache http client.
@@ -28,9 +16,8 @@ import java.util.concurrent.Executors;
  * HttpClient is synchronous, so the implementation is straight-forward.
  * IHttpClient is asynchronous, so we need to be more clever about integrating it with the synchronous apache client.
  */
-public class HttpClientWrapper implements com.azure.core.http.HttpClient, IHttpClient {
+public class HttpClientWrapper implements HttpClient, IHttpClient {
 
-    private static final Executor EXECUTOR = Executors.newCachedThreadPool();
     private final HttpClient httpClient;
 
     public HttpClientWrapper(HttpClient httpClient) {
@@ -40,93 +27,7 @@ public class HttpClientWrapper implements com.azure.core.http.HttpClient, IHttpC
     // Implementation of the asynchronous IHttpClient
     @Override
     public Mono<HttpResponse> send(HttpRequest httpRequest) {
-        // Creating the apache request
-
-        HttpUriRequest request;
-        String uri = httpRequest.getUrl().toString();
-        switch (httpRequest.getHttpMethod()) {
-            case GET:
-                request = new HttpGet(uri);
-                break;
-            case POST:
-                request = new HttpPost(uri);
-                break;
-            case PUT:
-                request = new HttpPut(uri);
-                break;
-            case PATCH:
-                request = new HttpPatch(uri);
-                break;
-            case DELETE:
-                request = new HttpDelete(uri);
-                break;
-            case HEAD:
-                request = new HttpHead(uri);
-                break;
-            case OPTIONS:
-                request = new HttpOptions(uri);
-                break;
-            case TRACE:
-                request = new HttpTrace(uri);
-                break;
-            default:
-                throw new IllegalArgumentException("Unsupported HTTP method: " + httpRequest.getHttpMethod());
-        }
-        // Translating the headers
-
-        request.setHeaders(httpRequest.getHeaders().stream().filter(h -> isNotContentLength(h.getName())).map(h -> new BasicHeader(h.getName(), h.getValue()))
-                .toArray(Header[]::new));
-
-        // Setting the request's body/entity
-
-        // This empty operation
-        if (request instanceof HttpEntityEnclosingRequest) {
-            HttpEntityEnclosingRequest entityEnclosingRequest = (HttpEntityEnclosingRequest) request;
-            PipedOutputStream osPipe = new PipedOutputStream();
-            PipedInputStream isPipe = null;
-            try {
-                isPipe = new PipedInputStream(osPipe);
-            } catch (IOException ignored) {
-                // This should never happen
-            }
-
-            EXECUTOR.execute(() -> {
-                httpRequest.getBody().publishOn(Schedulers.boundedElastic()).map(buf -> {
-                    try {
-                        if (buf.hasArray()) {
-                            osPipe.write(buf.array(), buf.position(), buf.remaining());
-                        } else {
-                            byte[] bytes = new byte[buf.remaining()];
-                            buf.get(bytes);
-                            osPipe.write(bytes);
-                        }
-                    } catch (IOException e) {
-                        return false;
-                    }
-                    return true;
-                }).blockLast();
-                try {
-                    osPipe.close();
-                } catch (IOException e) {
-                    // This should never happen
-                }
-            });
-
-            String contentLength = httpRequest.getHeaders().getValue("Content-Length");
-            String contentType = httpRequest.getHeaders().getValue("Content-Type");
-            entityEnclosingRequest.setEntity(new InputStreamEntity(isPipe, contentLength == null ? -1 : Long.parseLong(contentLength),
-                    contentType == null ? null : ContentType.parse(contentType)));
-        }
-
-        // The types of the Monos are different, but we ignore the results anyway (since we only care about the input stream) so this is fine.
-        return Mono.create(monoSink -> {
-            try {
-                org.apache.http.HttpResponse response = httpClient.execute(request);
-                monoSink.success(new HttpResponseWrapper(httpRequest, response));
-            } catch (IOException e) {
-                monoSink.error(e);
-            }
-        });
+        return httpClient.send(httpRequest);
     }
 
     private static boolean isNotContentLength(String name) {
@@ -135,34 +36,32 @@ public class HttpClientWrapper implements com.azure.core.http.HttpClient, IHttpC
 
     // Implementation of the synchronous HttpClient
     @Override
-    public IHttpResponse send(com.microsoft.aad.msal4j.HttpRequest httpRequest) throws Exception {
-        HttpUriRequest request;
-
-        // Translating the request
-
-        String uri = httpRequest.url().toString();
+    public IHttpResponse send(com.microsoft.aad.msal4j.HttpRequest httpRequest) {
+        HttpMethod method;
         switch (httpRequest.httpMethod()) {
             case GET:
-                request = new HttpGet(uri);
+                method = HttpMethod.GET;
                 break;
             case POST:
-                request = new HttpPost(uri);
+                method = HttpMethod.POST;
                 break;
             default:
                 throw new IllegalArgumentException("Unsupported HTTP method: " + httpRequest.httpMethod());
         }
-        // Translating the headers
 
-        request.setHeaders(httpRequest.headers().entrySet().stream().filter(h -> isNotContentLength(h.getKey()))
-                .map(h -> new BasicHeader(h.getKey(), h.getValue())).toArray(Header[]::new));
-
-        // Setting the request's body/entity
-        if (request instanceof HttpEntityEnclosingRequest) {
-            HttpEntityEnclosingRequest entityEnclosingRequest = (HttpEntityEnclosingRequest) request;
-            String body = httpRequest.body();
-            entityEnclosingRequest.setEntity(new ByteArrayEntity(body.getBytes(StandardCharsets.UTF_8)));
+        // Generate an azure core HttpRequest from the existing msal4j HttpRequest
+        HttpRequest request = new HttpRequest(method, httpRequest.url(), new HttpHeaders(httpRequest.headers()));
+        if (!CoreUtils.isNullOrEmpty(httpRequest.body())) {
+            request.setBody(httpRequest.body());
         }
 
-        return new HttpResponseWrapper(httpClient.execute(request));
+        try (HttpResponse response = httpClient.sendSync(request, Context.NONE)) {
+            com.microsoft.aad.msal4j.HttpResponse msalResponse = new com.microsoft.aad.msal4j.HttpResponse();
+            msalResponse.statusCode(response.getStatusCode());
+            msalResponse.body(response.getBodyAsBinaryData().toString());
+            msalResponse.addHeaders(response.getHeaders().stream().collect(Collectors.toMap(HttpHeader::getName,
+                    HttpHeader::getValuesList)));
+            return msalResponse;
+        }
     }
 }