+++ data/src/main/java/com/microsoft/azure/kusto/data/http/CloseParentResourcesStream.java
@@ -3,7 +3,8 @@
 
 package com.microsoft.azure.kusto.data.http;
 
-import org.apache.http.client.methods.CloseableHttpResponse;
+import com.azure.core.http.HttpResponse;
+import com.microsoft.azure.kusto.data.Utils;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -16,10 +17,10 @@ import java.io.InputStream;
  */
 public class CloseParentResourcesStream extends InputStream {
     private final InputStream innerStream;
-    private final CloseableHttpResponse httpResponse;
+    private final HttpResponse httpResponse;
 
-    public CloseParentResourcesStream(CloseableHttpResponse httpResponse) throws IOException {
-        this.innerStream = httpResponse.getEntity().getContent();
+    public CloseParentResourcesStream(HttpResponse httpResponse) throws IOException {
+        this.innerStream = Utils.getResponseAsStream(httpResponse);
         this.httpResponse = httpResponse;
     }