+++ data/src/main/java/com/microsoft/azure/kusto/data/http/HttpClientFactory.java
@@ -1,109 +1,70 @@
 package com.microsoft.azure.kusto.data.http;
 
-import com.microsoft.azure.kusto.data.HttpClientProperties;
-import org.apache.http.HeaderElement;
-import org.apache.http.HeaderElementIterator;
-import org.apache.http.HttpResponse;
-import org.apache.http.conn.ConnectionKeepAliveStrategy;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.message.BasicHeaderElementIterator;
-import org.apache.http.protocol.HTTP;
-import org.apache.http.protocol.HttpContext;
+import com.azure.core.http.HttpClient;
+import com.azure.core.http.HttpHeaderName;
+import com.azure.core.util.Header;
+import com.azure.core.util.HttpClientOptions;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.net.ssl.SSLContext;
-import java.security.NoSuchAlgorithmException;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
+import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * A singleton factory of HTTP clients.
+ * A static factory for HTTP clients.
  */
 public class HttpClientFactory {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(HttpClientFactory.class);
 
     /**
-     * Creates a new Apache HTTP client.
+     * Creates a new HTTP client.
      *
      * @param properties custom HTTP client properties
-     * @return a new Apache HTTP client
+     * @return a new HTTP client
      */
-    public static CloseableHttpClient create(HttpClientProperties properties) {
-        LOGGER.info("Creating new CloseableHttpClient client");
+    public static HttpClient create(HttpClientProperties properties) {
+        LOGGER.info("Creating new HTTP Client");
+        HttpClientOptions options = new HttpClientOptions();
+
+        // If all properties are null, create with default client options
         if (properties == null) {
-            properties = HttpClientProperties.builder().build();
+            options.setResponseTimeout(Duration.ofMinutes(10));
+            return HttpClient.createDefault(options);
         }
 
-        final HttpClientBuilder httpClientBuilder = HttpClientBuilder.create()
-                .useSystemProperties()
-                .setMaxConnTotal(properties.maxConnectionTotal())
-                .setMaxConnPerRoute(properties.maxConnectionRoute())
-                .evictExpiredConnections()
-                .evictIdleConnections(properties.maxIdleTime(), TimeUnit.SECONDS)
-                .disableRedirectHandling();
+        // MS Docs indicate that all setters handle nulls so even if these values are null everything should "just work"
+        // Note that the first discovered HttpClientProvider class is loaded. HttpClientProviders can be swapped in or out
+        // by simply
+        // Docs: https://learn.microsoft.com/en-us/java/api/com.azure.core.util.httpclientoptions?view=azure-java-stable
 
+        options.setMaximumConnectionPoolSize(properties.maxConnectionTotal());
+        options.setConnectionIdleTimeout(Duration.ofSeconds(properties.maxIdleTime()));
+        options.setResponseTimeout(properties.timeout());
+        options.setHttpClientProvider(properties.provider());
+
+        // Set Keep-Alive headers if they were requested.
+        // NOTE: Servers are not obligated to honor client requested Keep-Alive values
         if (properties.isKeepAlive()) {
-            final ConnectionKeepAliveStrategy keepAliveStrategy = new CustomConnectionKeepAliveStrategy(properties.maxKeepAliveTime());
-            httpClientBuilder.setKeepAliveStrategy(keepAliveStrategy);
-        }
+            Header keepAlive = new Header(HttpHeaderName.CONNECTION.getCaseSensitiveName(), "Keep-Alive");
+            // Keep-Alive is Non-standard from the client so core does not have an enum for it
+            Header keepAliveTimeout = new Header("Keep-Alive", "timeout=" + properties.maxKeepAliveTime());
 
-        if (properties.getPlanner() != null) {
-            httpClientBuilder.setRoutePlanner(properties.getPlanner());
+            List<Header> headers = new ArrayList<>();
+            headers.add(keepAlive);
+            headers.add(keepAliveTimeout);
+
+            options.setHeaders(headers);
         }
 
         if (properties.getProxy() != null) {
-            httpClientBuilder.setProxy(properties.getProxy());
+            options.setProxyOptions(properties.getProxy());
         }
 
-        if (properties.supportedProtocols() != null) {
+        // Todo: Is the per route connection maximum needed anymore?
 
-            try {
-                httpClientBuilder.setSSLSocketFactory(new SSLConnectionSocketFactory(SSLContext.getDefault(), properties.supportedProtocols(), null,
-                        SSLConnectionSocketFactory.getDefaultHostnameVerifier()));
-            } catch (NoSuchAlgorithmException e) {
-                LOGGER.error("Failed to set supported protocols", e);
-            }
-        }
-
-        return httpClientBuilder.build();
-    }
-
-    /**
-     * A custom connection keep-alive strategy that uses the server instructions set in the {@code Keep-Alive}
-     * response header; if the response doesn't contain a {@code Keep-Alive} header, the client will use a configurable
-     * keep-alive period.
-     */
-    static class CustomConnectionKeepAliveStrategy implements ConnectionKeepAliveStrategy {
-
-        private final int defaultKeepAlive;
-
-        /**
-         * The default keep-alive time.
-         *
-         * @param defaultKeepAlive the keep-alive time expressed in seconds
-         */
-        CustomConnectionKeepAliveStrategy(int defaultKeepAlive) {
-            this.defaultKeepAlive = defaultKeepAlive;
-        }
-
-        @Override
-        public long getKeepAliveDuration(HttpResponse httpResponse, HttpContext httpContext) {
-            // honor 'keep-alive' header
-            HeaderElementIterator it = new BasicHeaderElementIterator(httpResponse.headerIterator(HTTP.CONN_KEEP_ALIVE));
-            while (it.hasNext()) {
-                HeaderElement he = it.nextElement();
-                String param = he.getName();
-                String value = he.getValue();
-                if (value != null && param.equalsIgnoreCase("timeout")) {
-                    return Long.parseLong(value) * 1000;
-                }
-            }
-            // otherwise keep alive for default seconds
-            return defaultKeepAlive * 1000L;
-        }
+        return HttpClient.createDefault(options);
     }
 }