+++ data/src/main/java/com/microsoft/azure/kusto/data/instrumentation/MonitoredActivity.java
@@ -2,6 +2,7 @@ package com.microsoft.azure.kusto.data.instrumentation;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.Callable;
 
 public class MonitoredActivity {
     public static void invoke(Runnable runnable, String nameOfSpan) {
@@ -11,6 +12,16 @@ public class MonitoredActivity {
     public static void invoke(Runnable runnable, String nameOfSpan, Map<String, String> attributes) {
         try (Tracer.Span span = Tracer.startSpan(nameOfSpan, attributes)) {
             runnable.run();
+        }
+    }
+
+    public static <T> T invoke(SupplierNoException<T> supplier, String nameOfSpan) {
+        return invoke(supplier, nameOfSpan, new HashMap<>());
+    }
+
+    public static <T> T invoke(SupplierNoException<T> supplier, String nameOfSpan, Map<String, String> attributes) {
+        try (Tracer.Span span = Tracer.startSpan(nameOfSpan, attributes)) {
+            return supplier.get();
         }
     }