+++ samples/src/main/java/Query.java
@@ -4,7 +4,7 @@
 import com.microsoft.azure.kusto.data.Client;
 import com.microsoft.azure.kusto.data.ClientFactory;
 import com.microsoft.azure.kusto.data.ClientRequestProperties;
-import com.microsoft.azure.kusto.data.HttpClientProperties;
+import com.microsoft.azure.kusto.data.http.HttpClientProperties;
 import com.microsoft.azure.kusto.data.KustoOperationResult;
 import com.microsoft.azure.kusto.data.KustoResultSetTable;
 import com.microsoft.azure.kusto.data.auth.ConnectionStringBuilder;
@@ -30,7 +30,7 @@ public class Query {
 
             Client client = ClientFactory.createClient(csb, properties);
 
-            KustoOperationResult results = client.execute(".show version");
+            KustoOperationResult results = client.executeQuery(".show version");
             KustoResultSetTable mainTableResult = results.getPrimaryResults();
             System.out.printf("Kusto sent back %s rows.%n", mainTableResult.count());
 
@@ -43,7 +43,7 @@ public class Query {
             ClientRequestProperties clientRequestProperties = new ClientRequestProperties();
             clientRequestProperties.setTimeoutInMilliSec(TimeUnit.MINUTES.toMillis(1));
 
-            results = client.execute(System.getProperty("dbName"), System.getProperty("query"), clientRequestProperties);
+            results = client.executeQuery(System.getProperty("dbName"), System.getProperty("query"), clientRequestProperties);
         } catch (Exception e) {
             e.printStackTrace();
         }